add user reset and apply setup
