+++ frontend/src/components/Initiator.js
@@ -4,10 +4,12 @@ import { useDispatch } from "react-redux";
 import { useAuth } from "utils/hooks/useAuth";
 import { fetchUser } from "features/userSlice";
 import { fetchOptions } from "features/optionsSlice";
+import { getProfileId, getRole } from "utils/auth.service";
 
 function Initiator() {
   const { i18n } = useTranslation();
-  const { profileId, role } = useAuth();
+  const profileId = getProfileId();
+  const role = getRole();
   const dispatch = useDispatch();
 
   useEffect(() => {
