+++ frontend/src/components/PrivateRoute.js
@@ -1,14 +1,13 @@
 import React from "react";
 import { Route, Redirect } from "react-router-dom";
-import { useAuth } from "utils/hooks/useAuth";
+import { getUserIdToken } from "utils/auth.service";
 
 function PrivateRoute({ children, ...rest }) {
-  const { profileId } = useAuth();
   return (
     <Route
       {...rest}
       render={({ location }) =>
-        profileId ? (
+        getUserIdToken() ? (
           children
         ) : (
           <Redirect
