+++ frontend/src/components/css/Apply.scss
@@ -1,42 +1,3 @@
-.container2 {
-  position: relative;
-  width: 100%;
-  background: #ffdb6f;
-  height: 100vh;
-  overflow: auto;
-  // margin-bottom: 50px;
-}
-.home-header {
-  font-family: "Abel";
-  font-style: normal;
-  font-weight: 400;
-  font-size: 48px;
-  line-height: 61px;
-  color: #ffffff;
-  text-align: center;
-  position: absolute;
-  left: 0;
-  right: 0;
-  top: 36px;
-}
-
-.home-header span {
-  color: #644a0a;
-}
-.home-text {
-  height: 68px;
-  text-align: center;
-  font-family: "Abel";
-  font-style: normal;
-  font-weight: 400;
-  font-size: 48px;
-  line-height: 61px;
-  color: #ffffff;
-  position: absolute;
-  left: 0;
-  right: 0;
-  top: 109px;
-}
 .emailPart {
   position: absolute;
   width: 80%;
