+++ frontend/src/components/pages/PublicProfile.js
@@ -1,19 +1,24 @@
 import React, { useEffect, useState } from "react";
 import { UserOutlined } from "@ant-design/icons";
 import { Avatar } from "antd";
+import { withRouter } from "react-router-dom";
 
 import ProfileContent from "../ProfileContent";
 import ProfileVideos from "../ProfileVideos";
-import { fetchAccountById, getMenteePrivateStatus } from "../../utils/api";
+import { fetchAccountById } from "../../utils/api";
 
 import "../css/PublicProfile.scss";
 import { ACCOUNT_TYPE } from "utils/consts";
 import MenteeVideo from "components/MenteeVideo";
 
-function PublicProfile({ accountType, id }) {
+function PublicProfile({ match }) {
+  const accountType = match.params.type;
+  const id = match.params.id;
   const [account, setAccount] = useState({});
   const [updateContent, setUpdateContent] = useState(false);
   const [isMentor, setIsMentor] = useState(accountType == ACCOUNT_TYPE.MENTOR);
+
+  console.log(match);
 
   useEffect(() => {
     async function getAccount() {
@@ -64,4 +69,4 @@ function PublicProfile({ accountType, id }) {
   );
 }
 
-export default PublicProfile;
+export default withRouter(PublicProfile);
