+++ frontend/src/index.js
@@ -11,15 +11,19 @@ import "moment/locale/ar";
 import "moment/locale/fa";
 import "moment/locale/pt";
 import i18n from "utils/i18n";
+import { ProvideAuth } from "utils/hooks/useAuth";
+import { ConfigProvider } from "antd";
 
 moment.locale(i18n.language);
 
 ReactDOM.render(
   <React.StrictMode>
     <Provider store={store}>
-      <Suspense>
-        <App />
-      </Suspense>
+      <ProvideAuth>
+        <Suspense>
+          <App />
+        </Suspense>
+      </ProvideAuth>
     </Provider>
   </React.StrictMode>,
   document.getElementById("root")
