+++ frontend/src/utils/consts.js
@@ -40,17 +40,18 @@ export const ACCOUNT_TYPE = {
 };
 
 export const ACCOUNT_TYPE_LABELS = {
-  [ACCOUNT_TYPE.ADMIN]: "Admin",
-  [ACCOUNT_TYPE.MENTOR]: "Mentor",
-  [ACCOUNT_TYPE.MENTEE]: "Mentee",
-  [ACCOUNT_TYPE.PARTNER]: "Partner",
-  [ACCOUNT_TYPE.GUEST]: "Guest",
+  [ACCOUNT_TYPE.ADMIN]: "admin",
+  [ACCOUNT_TYPE.MENTOR]: "mentor",
+  [ACCOUNT_TYPE.MENTEE]: "mentee",
+  [ACCOUNT_TYPE.PARTNER]: "partner",
+  [ACCOUNT_TYPE.GUEST]: "guest",
 };
 
 export const REDIRECTS = {
   [ACCOUNT_TYPE.MENTOR]: "/appointments",
   [ACCOUNT_TYPE.MENTEE]: "/mentee-appointments",
   [ACCOUNT_TYPE.PARTNER]: "/partner-gallery",
+  [ACCOUNT_TYPE.ADMIN]: "/account-data",
 };
 
 export const PLURAL_TYPE = {
